move newsnr-like functions out of events.py into new module
fixing initial bugs in ranking.py

small fix to ranking.py
