+++ pycbc/events/__init__.py
@@ -5,4 +5,6 @@ This packages contains modules for clustering events
 from .events import *
 from .veto import *
 from .coinc import *
+from .triggers import *
+from .ranking import *
