+++ pycbc/events/ranking.py
@@ -1,0 +1,62 @@
+import numpy
+
+
+def effsnr(snr, reduced_x2, fac=250.):
+    """Calculate the effective SNR statistic. See (S5y1 paper) for definition.
+    """
+    snr = numpy.array(snr, ndmin=1, dtype=numpy.float64)
+    rchisq = numpy.array(reduced_x2, ndmin=1, dtype=numpy.float64)
+    esnr = snr / (1 + snr ** 2 / fac) ** 0.25 / rchisq ** 0.25
+
+    # If snr input is float, return a float. Otherwise return numpy array.
+    if hasattr(snr, '__len__'):
+        return esnr
+    else:
+        return esnr[0]
+
+def newsnr(snr, reduced_x2, q=6., n=2.):
+    """Calculate the re-weighted SNR statistic ('newSNR') from given SNR and
+    reduced chi-squared values. See http://arxiv.org/abs/1208.3491 for
+    definition. Previous implementation in glue/ligolw/lsctables.py
+    """
+    nsnr = numpy.array(snr, ndmin=1, dtype=numpy.float64)
+    reduced_x2 = numpy.array(reduced_x2, ndmin=1, dtype=numpy.float64)
+
+    # newsnr is only different from snr if reduced chisq > 1
+    ind = numpy.where(reduced_x2 > 1.)[0]
+    nsnr[ind] *= (0.5 * (1. + reduced_x2[ind] ** (q/n))) ** (-1./q)
+
+    # If snr input is float, return a float. Otherwise return numpy array.
+    if hasattr(snr, '__len__'):
+        return nsnr
+    else:
+        return nsnr[0]
+
+def newsnr_sgveto(snr, bchisq, sgchisq):
+    """ Combined SNR derived from NewSNR and Sine-Gaussian Chisq"""
+    nsnr = numpy.array(newsnr(snr, bchisq), ndmin=1)
+    sgchisq = numpy.array(sgchisq, ndmin=1)
+    t = numpy.array(sgchisq > 4, ndmin=1)
+    if len(t):
+        nsnr[t] = nsnr[t] / (sgchisq[t] / 4.0) ** 0.5
+
+    # If snr input is float, return a float. Otherwise return numpy array.
+    if hasattr(snr, '__len__'):
+        return nsnr
+    else:
+        return nsnr[0]
+
+def newsnr_sgveto_psdvar(snr, bchisq, sgchisq, psd_var_val):
+    """ Combined SNR derived from NewSNR, Sine-Gaussian Chisq and PSD
+    variation statistic """
+    nsnr = numpy.array(newsnr_sgveto(snr, bchisq, sgchisq), ndmin=1)
+    psd_var_val = numpy.array(psd_var_val, ndmin=1)
+    lgc = psd_var_val >= 1.8
+    nsnr[lgc] = nsnr[lgc] / numpy.sqrt(psd_var_val[lgc])
+
+    # If snr input is float, return a float. Otherwise return numpy array.
+    if hasattr(snr, '__len__'):
+        return nsnr
+    else:
+        return nsnr[0]
+
